test: Controller와 RestController를 분리
