+++ src/test/java/roomescape/controller/ReservationRestControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import static org.hamcrest.Matchers.is;
 
@@ -13,25 +13,11 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ReservationControllerTest {
-
-    @Test
-    @DisplayName("/admin으로 요청시 메인 페이지 응답")
-    void getMainPage() {
-        RestAssured.given().log().all()
-                .when().get("/admin")
-                .then().log().all()
-                .statusCode(200);
-    }
+public class ReservationRestControllerTest {
 
     @Test
     @DisplayName("/admin/reservation 요청 시 예약 관리 페이지 응답")
     void readReservation() {
-        RestAssured.given().log().all()
-                .when().get("/admin/reservation")
-                .then().log().all()
-                .statusCode(200);
-
         RestAssured.given().log().all()
                 .when().get("/reservations")
                 .then().log().all()
@@ -42,14 +28,9 @@ public class ReservationControllerTest {
     @Test
     @DisplayName("예약 관리 페이지 내에서 예약 추가")
     void postReservation() {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", "2023-08-05");
-        params.put("time", "15:40");
-
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
-                .body(params)
+                .body(getTestParamsWithReservation())
                 .when().post("/reservations")
                 .then().log().all()
                 .statusCode(200)
@@ -65,14 +46,9 @@ public class ReservationControllerTest {
     @Test
     @DisplayName("예약 관리 페이지 내에서 예약 삭제")
     void deleteReservation() {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", "2023-08-05");
-        params.put("time", "15:40");
-
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
-                .body(params)
+                .body(getTestParamsWithReservation())
                 .when().post("/reservations")
                 .then().log().all()
                 .statusCode(200)
@@ -83,7 +59,7 @@ public class ReservationControllerTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(1));
-
+        
         RestAssured.given().log().all()
                 .when().delete("/reservations/1")
                 .then().log().all()
@@ -94,5 +70,13 @@ public class ReservationControllerTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(0));
+    }
+
+    private Map<String, String> getTestParamsWithReservation() {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("time", "15:40");
+        return params;
     }
 }
